pyproject.toml
[project]
name = "toonslate-backend"
version = "0.1.0"
description = "Toonslate backend - 웹툰 자동 번역 서비스"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.128.0",
    "gradio-client>=1.0.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "celery-types>=0.24.0",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "ruff>=0.14.14",
]
test = [
    "fakeredis[lua]>=2.33.0",
    "httpx>=0.28.1",
    "numpy>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]
ai = [
    "google-genai>=1.0.0",
    "opencv-python-headless>=4.10.0",
    "pillow>=10.0.0",
    "replicate>=1.0.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RET"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
